agri-microservice/disease/requirements.txt
# fastapi==0.110.0
# uvicorn==0.29.0
# python-multipart==0.0.9
# numpy>=1.26.0
# pillow>=10.3.0
# tensorflow>=2.16.1
# pydantic>=2.7.0



fastapi
uvicorn
python-multipart
numpy
pillow
tensorflow
pydantic